requirements.txt
gradio
pandas
python-dotenv
google-generativeai
Pillow
requests 